# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: agent-wire-docling
  version: 0.1.0
  description: |
    HTTP API for the agent-wire-docling prototype. Single source of truth for
    backend ↔ frontend. Frontend client is codegen'd from this file via
    openapi-typescript. All agents code against this; any mid-build gap is
    escalated, not silently edited.
servers:
  - url: http://localhost:8000

# ─────────────────────────────────────────────────────────────────────────────
# Tags
...
paths:
...
  /scan:
    post:
      tags: [scan]
      summary: Walk a folder, detect formats, cluster into strata
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScanRequest' }
      responses:
        '200':
          description: Scan result with stratum breakdown
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScanResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /strata/sample:
    post:
      tags: [scan]
      summary: Seeded, deterministic stratified sample pick
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SampleRequest' }
      responses:
        '200':
          description: Sampled docs grouped by stratum
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SampleResult' }
  /filemap:
    get:
      tags: [scan]
      summary: Read a folder's filemap (Level B)
      parameters:
        - in: query
          name: folder
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Filemap dict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Filemap' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [scan]
      summary: Merge user-owned fields into a folder's filemap
      parameters:
        - in: query
          name: folder
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FilemapPatchRequest' }
      responses:
        '200':
          description: Updated filemap
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Filemap' }
  /filetree:
    get:
      tags: [scan]
      summary: Recursive filetree view for UI rendering
      parameters:
        - in: query
          name: root
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Filetree node
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FiletreeNode' }
...
components:
  responses:
    BadRequest:
      description: Request validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        detail: { type: string }
...
    ScanRequest:
      type: object
      required: [folder]
      properties:
        folder: { type: string, description: 'Absolute local path' }
        follow_symlinks: { type: boolean, default: false }
        max_files: { type: integer, default: 50000 }
    ScanResult:
      type: object
      required: [folder, total_files, strata, skipped, scan_id]
      properties:
        scan_id: { type: string }
        folder: { type: string }
        total_files: { type: integer }
        strata:
          type: array
          items: { $ref: '#/components/schemas/Stratum' }
        skipped:
          type: array
          items:
            type: object
            required: [path, reason]
            properties:
              path: { type: string }
              reason: { type: string, description: 'e.g. "tier3_audio", "symlink_denied", "unreadable"' }
    Stratum:
      type: object
      required: [name, size, sample_size_hint, example_paths]
      properties:
        name: { type: string, description: 'e.g. pdf-native-11-50, docx, xlsx' }
        size: { type: integer }
        sample_size_hint: { type: integer, description: 'min(N, size) — exhaustive if size ≤ 6' }
        exhaustive: { type: boolean, description: 'true when size ≤ 6' }
        example_paths:
          type: array
          items: { type: string }
          description: 'First few filenames, for UI preview'
    SampleRequest:
      type: object
      required: [scan_id, n]
      properties:
        scan_id: { type: string }
        n: { type: integer, default: 5 }
        seed: { type: integer, description: 'Optional; deterministic resampling' }
        exclude_hashes:
          type: array
          items: { type: string }
          description: 'Already-approved/rejected hashes to skip'
    SampleResult:
      type: object
      required: [seed, strata]
      properties:
        seed: { type: integer }
        strata:
          type: array
          items:
            type: object
            required: [name, docs]
            properties:
              name: { type: string }
              docs:
                type: array
                items: { $ref: '#/components/schemas/DocRef' }
    DocRef:
      type: object
      required: [source_sha256, source_path, source_format]
      properties:
        source_sha256: { type: string }
        source_path: { type: string }
        source_format: { type: string, description: 'pdf, docx, xlsx, pptx, html, txt, md, latex' }
        size_bytes: { type: integer }
        page_count: { type: integer, nullable: true }

    # ── Convert / pipeline
...
    Filemap:
      type: object
      description: .understanding/folder.yaml parsed as JSON
      additionalProperties: true
    FilemapFileUpdate:
      type: object
      required: [path]
      properties:
        path: { type: string }
        user_included: { type: boolean, nullable: true }
        user_content_type: { type: string, nullable: true }
        user_notes: { type: string, nullable: true }
    FilemapPatchRequest:
      type: object
      properties:
        files:
          type: array
          items: { $ref: '#/components/schemas/FilemapFileUpdate' }
        defaults:
          type: object
          nullable: true
          additionalProperties: true
    FiletreeNode:
      type: object
      properties:
        path: { type: string }
        folder_relative: { type: string }
        counts:
          type: object
          properties:
            included: { type: integer }
            pending: { type: integer }
            excluded: { type: integer }
            total: { type: integer }
        filemap: { type: boolean }
        children:
          type: array
          items: { $ref: '#/components/schemas/FiletreeNode' }
...